# Conflicts:
#	continew-common/src/main/java/top/continew/admin/common/context/UserContext.java
#	continew-module-system/src/main/java/top/continew/admin/auth/service/impl/LoginServiceImpl.java
#	continew-webapi/src/main/java/top/continew/admin/controller/auth/SocialAuthController.java
#	continew-webapi/src/main/java/top/continew/admin/controller/system/DictItemController.java
#	continew-webapi/src/main/java/top/continew/admin/controller/system/FileController.java
#	continew-webapi/src/main/resources/config/application-dev.yml
#	continew-webapi/src/main/resources/config/application-prod.yml

+++ continew-module-system/src/main/java/top/continew/admin/system/model/entity/ClientDO.java
@@ -21,15 +21,16 @@ import com.baomidou.mybatisplus.annotation.TableName;
 import com.baomidou.mybatisplus.extension.handlers.JacksonTypeHandler;
 import lombok.Data;
 import top.continew.admin.common.enums.DisEnableStatusEnum;
-import top.continew.starter.extension.crud.model.entity.BaseDO;
+import top.continew.admin.common.model.entity.BaseDO;
 
 import java.io.Serial;
 import java.util.List;
 
 /**
- * 客户端管理实体
+ * 客户端实体
  *
- * @author MoChou
+ * @author KAI
+ * @author Charles7c
  * @since 2024/12/03 16:04
  */
 @Data
@@ -40,12 +41,12 @@ public class ClientDO extends BaseDO {
     private static final long serialVersionUID = 1L;
 
     /**
-     * 客户端ID
+     * 客户端 ID
      */
     private String clientId;
 
     /**
-     * 客户端Key
+     * 客户端 Key
      */
     private String clientKey;
 
@@ -66,17 +67,17 @@ public class ClientDO extends BaseDO {
     private String clientType;
 
     /**
-     * Token最低活跃频率（-1为不限制）
+     * Token 最低活跃频率（单位：秒，-1：不限制，永不冻结）
      */
-    private Integer activeTimeout;
+    private Long activeTimeout;
 
     /**
-     * Token有效期（默认30天，单位：秒）
+     * Token 有效期（单位：秒，-1：永不过期）
      */
-    private Integer timeout;
+    private Long timeout;
 
     /**
-     * 状态（1：启用；2：禁用）
+     * 状态
      */
     private DisEnableStatusEnum status;
 }
